astro-backend/requirements.txt
flask
flask_cors
flask_sqlalchemy
flask_migrate
flask_wtf
flask_login
flask_admin
jwt